Refactor interpolation code a bit, renaming some outdated material. Breakout interpolation class to the interpolation directory. Add in a singular track solver. Update interpolation notebook.
