+++ spyral/core/track_generator.py
@@ -1,5 +1,5 @@
 from .constants import MEV_2_JOULE, MEV_2_KG, C, E_CHARGE
-from ..interpolate import BilinearInterpolator, LinearInterpolator
+from ..interpolate import LinearInterpolator
 
 from spyral_utils.nuclear import NucleusData
 from spyral_utils.nuclear.target import GasTarget
@@ -11,19 +11,19 @@ from scipy.integrate import solve_ivp
 from pathlib import Path
 import json
 
-from numba import float64, int32
-from numba.types import string, ListType
-from numba.typed import List
-from numba.experimental import jitclass
 
-TIME_WINDOW: float = 1.0e-6  # 1us window
+COARSE_TIME_WINDOW: float = 1.0e-6  # 1us window
 KE_LIMIT = 0.1
 DEG2RAD: float = np.pi / 180.0
 
 
 @dataclass
-class InterpolatorParameters:
-    """Wrapper for the general parameters required to generate an interpolation scheme
+class MeshParameters:
+    """Wrapper for the general parameters required to generate a mesh of ODE solutions
+
+    A mesh for Spyral is a 4 dimensional array of solutions to the equations of motion (also
+    called track or trajectory) through the AT-TPC detector. These parameters describe
+    the shape of the mesh as well as the physical properties used to generate it.
 
     Attributes
     ----------
@@ -38,17 +38,17 @@ class InterpolatorParameters:
     n_time_steps: int
         The number of ODE solver timesteps
     ke_min: float
-        The minimum kinetic energy of the interpolation scheme in MeV
+        The minimum kinetic energy of the mesh in MeV
     ke_max: float
-        The maximum kinetic energy of the interpolation scheme in MeV
+        The maximum kinetic energy of the mesh in MeV
     ke_bins: int
-        The number of kinetic energy bins in the interpolation scheme
+        The number of kinetic energy bins in the mesh
     polar_min: float
-        The minimum polar angle of the interpolation scheme in degrees
+        The minimum polar angle of the mesh in degrees
     polar_max: float
-        The maximum polar angle of the interpolation scheme in degrees
+        The maximum polar angle of the mesh in degrees
     polar_bins: int
-        The number of polar angle bins in the interpolation scheme
+        The number of polar angle bins in the mesh
 
     Methods
     -------
@@ -93,6 +93,32 @@ class InterpolatorParameters:
             },
             indent=4,
         )
+
+
+def check_mesh_needs_generation(track_path: Path, params: MeshParameters) -> bool:
+    """Check if track mesh meta data matches or if track mesh doesn't exist
+
+    Parameters
+    ----------
+    track_path: Path
+        Path to the track mesh data
+    params: MeshParamters
+        The parameters for this track mesh
+
+    Returns
+    -------
+    bool
+        Returns True if the mesh needs to be generated
+    """
+    if track_path.exists():
+        meta_path = track_path.parents[0] / f"{track_path.stem}.json"
+        if not meta_path.exists():
+            return True
+        with open(meta_path, "r") as meta_file:
+            meta_str = meta_file.read()
+            return params.serialize_json() != meta_str
+    else:
+        return True
 
 
 # State = [x, y, z, vx, vy, vz]
@@ -275,39 +301,20 @@ def rho_bound_condition(
         this function returns zero the termination condition has been reached.
 
     """
-    return np.linalg.norm(state[:2]) - 0.332
+    return float(np.linalg.norm(state[:2])) - 0.332
 
 
-def check_tracks_need_generation(
-    track_path: Path, params: InterpolatorParameters
-) -> bool:
-    """Check if track meta data matches or if tracks don't exist
+def generate_track_mesh(params: MeshParameters, track_path: Path):
+    """Generate a mesh of tracks given some parameters and write them to an npy file at a path.
 
-    Parameters
-    ----------
-    track_path: Path
-        Path to the track interpolation scheme
-    params: InterpolatorParameters
-        The parameters for this interpolation scheme
+    Creates a 4-dimensional array (mesh) of ODE solutions (also called tracks or trajectories) and
+    writes them to disk. The dimensions are as follows: time x initial energy x inital polar angle x position.
+    This mesh can then be used to interpolate and find an approximate solution for a given energy, polar angle.
 
-    Returns
-    -------
-    bool
-        Returns True if the interpolation scheme needs to be generated
-    """
-    if track_path.exists():
-        meta_path = track_path.parents[0] / f"{track_path.stem}.json"
-        if not meta_path.exists():
-            return True
-        with open(meta_path, "r") as meta_file:
-            meta_str = meta_file.read()
-            return params.serialize_json() != meta_str
-    else:
-        return True
-
-
-def generate_tracks(params: InterpolatorParameters, track_path: Path):
-    """Generate a set of tracks given some parameters and write them to an npy file at a path.
+    In many cases of AT-TPC analysis, it is diffcult to pre-estimate the time window needed to cover the full
+    trajectory. To compensate, Spyral will do a first pass over the mesh using a coarse 1us time window.
+    It will then estimate a finer time window based off of the results of this first pass, and then generate the
+    mesh.
 
     Parameters
     ----------
@@ -325,10 +332,11 @@ def generate_tracks(params: InterpolatorParameters, track_path: Path):
     with open(track_meta_path, "w") as metafile:
         metafile.write(params.serialize_json())
 
+    # time x (x, y, z, vx, vy, vz) x ke x polar
     data = np.zeros(
         shape=(params.n_time_steps, 3, params.ke_bins, params.polar_bins),
         dtype=np.float64,
-    )  # time x (x, y, z, vx, vy, vz) x ke x polar
+    )
 
     initial_state = np.zeros(6)
 
@@ -343,6 +351,8 @@ def generate_tracks(params: InterpolatorParameters, track_path: Path):
 
     longest_time = 0.0
 
+    # Set the conditions to be terminal for a given direction (approaching from positive or negative slope)
+    # See scipy docs for details
     stop_condition.terminal = True
     stop_condition.direction = -1.0
     z_bound_condition.terminal = True
@@ -371,7 +381,7 @@ def generate_tracks(params: InterpolatorParameters, track_path: Path):
             initial_state[5] = speed * math.cos(p)
             result = solve_ivp(
                 equation_of_motion,
-                (0.0, TIME_WINDOW),
+                (0.0, COARSE_TIME_WINDOW),
                 initial_state,
                 method="RK45",
                 events=[stop_condition, z_bound_condition, rho_bound_condition],
@@ -421,371 +431,106 @@ def generate_tracks(params: InterpolatorParameters, track_path: Path):
     np.save(track_path, data)
 
 
-# To use numba with a class we need to declare the types of all members of the class
-# and use the @jitclass decorator
-tinterp_spec = [
-    ("filepath", string),
-    ("particle_name", string),
-    ("gas_name", string),
-    ("bfield", float64),
-    ("efield", float64),
-    ("ke_min", float64),
-    ("ke_max", float64),
-    ("ke_bins", int32),
-    ("polar_min", float64),
-    ("polar_max", float64),
-    ("polar_bins", int32),
-    ("interpolators", ListType(BilinearInterpolator.class_type.instance_type)),
-]
-
-
-@jitclass(spec=tinterp_spec)
-class TrackInterpolator:
-    """Represents an interpolation scheme used to generate trajectories.
-
-    Solving ODE's can be expensive, so to save time pre-generate a range of solutions and then
-    interpolate on these solutions. TrackInterpolator uses bilinear interpolation to
-    interpolate on the energy and polar angle (reaction angle) of the trajectory.
-
-    We use numba to just-in-time compile these methods, which results in a dramatic speed up on the order
-    of a factor of 50.
-
-    Attriubtes
-    ----------
-    file_path: str
-        The track save file
-    particle_name: str
-        The isotopic symbol of the ejectile
-    gas_name: str
-        The target gas name
-    bfield: float
-        The magnetic field magnitude in T
-    efield: float
-        The electric field magnitude in V/m
-    ke_min: float
-        The minimum kinetic energy of the interpolation scheme in MeV
-    ke_max: float
-        The maximum kinetic energy of the interpolation scheme in MeV
-    ke_bins: int
-        The number of kinetic energy bins in the interpolation scheme
-    polar_min: float
-      The minimum polar angle of the interpolation scheme in degrees
-    polar_max: float
-      The maximum polar angle of the interpolation scheme in degrees
-    polar_bins: int
-        The number of polar angle bins in the interpolation scheme
-    interpolators: ListType[BilinearInterpolator]
-        A list of BilinearInterpolators, one for each time step
-
-    Methods
-    -------
-    TrackInterpolator(
-        track_path: str,
-        interpolators: ListType[BilinearInterpolator],
-        particle_name: str,
-        gas_name: str,
-        bfield: float,
-        efield: float,
-        ke_min: float,
-        ke_max: float,
-        ke_bins: int,
-        polar_min: float,
-        polar_max: float,
-        polar_bins: int)
-        Construct a TrackInterpolator
-
-    get_interpolated_trajectory(vx: float, vy: float, vz: float, polar: float, azim: float, ke: float) -> LinearInterpolator | None
-        Given some initial state, get an interpolated trajectory
-
-    check_interpolator(
-        particle: str,
-        bfield: float,
-        efield: float,
-        target: str,
-        ke_min: float,
-        ke_max: float,
-        ke_bins: int,
-        polar_min: float,
-        polar_max: float,
-        polar_bins: int,
-    ) -> bool
-        Check if this interpolator matches the given values
-
-    check_values_in_range(ke: float, polar: float) -> bool
-        Check if the given ke, polar point is within the interpolation scheme
-    """
-
-    def __init__(
-        self,
-        track_path: str,
-        interpolators: ListType(BilinearInterpolator.class_type.instance_type),
-        particle_name: str,
-        gas_name: str,
-        bfield: float,
-        efield: float,
-        ke_min: float,
-        ke_max: float,
-        ke_bins: int,
-        polar_min: float,
-        polar_max: float,
-        polar_bins: int,
-    ):
-        """Construct a TrackInterpolator
-
-        Parameters
-        ----------
-        track_path: str
-            Path to an interpolation file
-        interpolators: ListType[BilinearInterpolator]
-            A set of BilinearInterpolators
-        particle_name: str
-            The isotopic symbol of the particle
-        gas_name: str
-            The gas target name
-        bfield: float
-            The magnetic field magnitude in T
-        efield: float
-            The electric field magnitude in V/m
-        ke_min: float
-            The minimum kinetic energy of the interpolation scheme in MeV
-        ke_max: float
-            The maximum kinetic energy of the interpolation scheme in MeV
-        ke_bins: int
-            The number of kinetic energy bins in the interpolation scheme
-        polar_min: float
-            The minimum polar angle of the interpolation scheme in degrees
-        polar_max: float
-            The maximum polar angle of the interpolation scheme in degrees
-        polar_bins: int
-            The number of polar angle bins in the interpolation scheme
-
-        Returns
-        -------
-        TrackInterpolator
-            An instance of the class
-        """
-        self.filepath = track_path
-        self.particle_name: str = particle_name
-        self.gas_name: str = gas_name
-        self.bfield: float = bfield
-        self.efield: float = efield
-        self.ke_min: float = ke_min
-        self.ke_max: float = ke_max
-        self.ke_bins: int = ke_bins
-        self.polar_min: float = polar_min
-        self.polar_max: float = polar_max
-        self.polar_bins: int = polar_bins
-
-        self.interpolators = interpolators
-
-    def get_interpolated_trajectory(
-        self, vx: float, vy: float, vz: float, polar: float, azim: float, ke: float
-    ) -> LinearInterpolator | None:
-        """Get an interpolated trajectory given some initial state.
-
-        Parameters
-        -------------
-        vx: float
-            Vertex x-coordinate in m
-        vy: float
-            Vertex y-coordinate in m
-        vz: float
-            Vertex z-coordinate in m
-        polar: float
-            Polar angle in radians
-        azim: float
-            azimuthal angle in radians
-        ke: float
-            Kinetic energy in MeV
-
-        Returns
-        -------
-        LinearInterpolator | None
-            Returns a LinearInterpolator, which interpolates the trajectory upon z for x,y or None when the algorithm fails
-        """
-
-        trajectory = np.zeros((len(self.interpolators), 3))
-        for idx, _ in enumerate(trajectory):
-            trajectory[idx] = self.interpolators[idx].interpolate(polar, ke)
-
-        # Rotate the trajectory in azimuthal (around z) to match data
-        z_rot = np.array(
-            [
-                [np.cos(azim), -np.sin(azim), 0.0],
-                [np.sin(azim), np.cos(azim), 0.0],
-                [0.0, 0.0, 1.0],
-            ]
-        )
-        trajectory = (z_rot @ trajectory.T).T
-        # Translate to vertex
-        trajectory[:, 0] += vx
-        trajectory[:, 1] += vy
-        trajectory[:, 2] += vz
-        # Trim stopped region
-        removal = np.full(len(trajectory), True)
-        previous_element = np.full(3, -1.0)
-        for idx, element in enumerate(trajectory):
-            if np.all(previous_element[:] == element[:]):
-                removal[idx] = False
-            previous_element = element
-
-        trajectory = trajectory[removal]
-        if len(trajectory) < 2:
-            return None
-
-        # Handle data > 90 deg. LinearInterpolator requires x data (z in our case) to be monotonically increasing
-        # So flip the array along axis 0 (flipud). Also trim data where sometimes particle starts going backward
-        # due to efield
-        if polar > np.pi * 0.5:
-            trajectory = np.flipud(trajectory)
-            mask = np.diff(np.ascontiguousarray(trajectory[:, 2])) > 0
-            mask = np.append(mask, True)
-            trajectory = trajectory[mask]
-
-        return LinearInterpolator(trajectory[:, 2], trajectory[:, :2].T)
-
-    def check_interpolator(
-        self,
-        particle: str,
-        bfield: float,
-        efield: float,
-        target: str,
-        ke_min: float,
-        ke_max: float,
-        ke_bins: int,
-        polar_min: float,
-        polar_max: float,
-        polar_bins: int,
-    ) -> bool:
-        """Check to see if this interpolator matches the given parameters
-
-        Parameters
-        ----------
-        particle: str
-            The isotopic symbol of the particle
-        bfield: float
-            The magnetic field magnitude in T
-        efield: float
-            The electric field magnitude in V/m
-        target: str
-            The gas target name
-        ke_min: float
-            The minimum kinetic energy of the interpolation scheme in MeV
-        ke_max: float
-            The maximum kinetic energy of the interpolation scheme in MeV
-        ke_bins: int
-            The number of kinetic energy bins in the interpolation scheme
-        polar_min: float
-            The minimum polar angle of the interpolation scheme in degrees
-        polar_max: float
-            The maximum polar angle of the interpolation scheme in degrees
-        polar_bins: int
-            The number of polar angle bins in the interpolation scheme
-
-        Returns
-        -------
-        bool
-            Returns true if the interpolator matches
-        """
-        is_valid = (
-            particle == self.particle_name
-            and bfield == self.bfield
-            and efield == self.efield
-            and target == self.gas_name
-            and ke_min == self.ke_min
-            and ke_max == self.ke_max
-            and ke_bins == self.ke_bins
-            and polar_min == self.polar_min
-            and polar_max == self.polar_max
-            and polar_bins == self.polar_bins
-        )
-
-        if is_valid:
-            return True
-        else:
-            return False
-
-    def check_values_in_range(self, ke: float, polar: float) -> bool:
-        """Check if these values of energy, angle are within the interpolation range
-
-        Parameters
-        ----------
-        ke: float
-            The kinetic energy to check in MeV
-        polar: float
-            The polar angle to check in radians
-
-        Returns
-        -------
-        bool
-            Returns true if they are within the interpolation range
-        """
-        polar_deg = polar / DEG2RAD
-        if (
-            ke > self.ke_max
-            or ke < self.ke_min
-            or polar_deg < self.polar_min
-            or polar_deg > self.polar_max
-        ):
-            return False
-        else:
-            return True
-
-
-def create_interpolator(track_path: Path) -> TrackInterpolator:
-    """Create a TrackInterpolator
-
-    This is a utility function wrapping the creation of a TrackInterpolator. Numba doesn't support h5py for
-    obvious reasons, so we need to retrieve the track data outside of the class and then manually pass it all
-    to the initialization.
+def generate_interpolated_track(
+    vx: float,
+    vy: float,
+    vz: float,
+    polar: float,
+    azim: float,
+    ke: float,
+    particle: NucleusData,
+    bfield: float,
+    efield: float,
+    target: GasTarget,
+    n_time_steps: int = 1000,
+) -> LinearInterpolator | None:
+    """Get a single interpolated trajectory given some initial state and system parameters
 
     Parameters
-    ----------
-    track_path: Path
-        Path to the track interpolation data
-
+    -------------
+    vx: float
+        Vertex x-coordinate in m
+    vy: float
+        Vertex y-coordinate in m
+    vz: float
+        Vertex z-coordinate in m
+    polar: float
+        Polar angle in radians
+    azim: float
+        azimuthal angle in radians
+    ke: float
+        Kinetic energy in MeV
+    particle: NucleusData
+        The particle of interest
+    bfield: float
+        The magnetic field in Tesla
+    efield: float
+        The electric field in V/m
+    target: GasTarget
+        The target material
+    n_time_steps: int
+        The number of timesteps used by the solver
     Returns
     -------
-    TrackInterpolator
-        The constructed interpolator
+    LinearInterpolator | None
+        Returns a LinearInterpolator, which interpolates the trajectory upon z for x,y or None when the algorithm fails
     """
-    track_meta_path = track_path.parents[0] / f"{track_path.stem}.json"
-    meta_dict: dict
-    with open(track_meta_path, "r") as metafile:
-        meta_dict = json.load(metafile)
-    data = np.load(track_path)
 
-    pmin_rad = meta_dict["polar_min"] * DEG2RAD
-    pmax_rad = meta_dict["polar_max"] * DEG2RAD
+    # Termination conditions
+    stop_condition.terminal = True
+    stop_condition.direction = -1.0
+    z_bound_condition.terminal = True
+    z_bound_condition.direction = 1.0
+    rho_bound_condition.terminal = True
+    rho_bound_condition.direction = 1.0
 
-    typed_interpolators = List()
-
-    [
-        typed_interpolators.append(
-            BilinearInterpolator(
-                pmin_rad,
-                pmax_rad,
-                meta_dict["polar_bins"],
-                meta_dict["ke_min"],
-                meta_dict["ke_max"],
-                meta_dict["ke_bins"],
-                time.T[:, :, :3],
-            )
-        )
-        for time in data
-    ]
-
-    return TrackInterpolator(
-        str(track_path),
-        typed_interpolators,
-        meta_dict["particle"],
-        meta_dict["gas"],
-        meta_dict["bfield"],
-        meta_dict["efield"],
-        meta_dict["ke_min"],
-        meta_dict["ke_max"],
-        meta_dict["ke_bins"],
-        meta_dict["polar_min"],
-        meta_dict["polar_max"],
-        meta_dict["polar_bins"],
+    # Setup initial state
+    momentum = math.sqrt(ke * (ke + 2.0 * particle.mass))
+    speed = momentum / particle.mass * C
+    initial_state = np.array(
+        [
+            vx,
+            vy,
+            vz,
+            speed * math.sin(polar) * math.cos(azim),
+            speed * math.sin(polar) * math.sin(azim),
+            speed * math.cos(polar),
+        ]
     )
+
+    # Solve coarsely to deterimine problem timescale
+    trajectory = solve_ivp(
+        equation_of_motion,
+        (0.0, COARSE_TIME_WINDOW),
+        initial_state,
+        method="RK45",
+        events=[stop_condition, z_bound_condition, rho_bound_condition],
+        args=(bfield, efield, target, particle),
+    )
+    longest_time = trajectory.t[-1]
+
+    # Solve with refined scale to produce best results
+    time_steps = np.linspace(0.0, longest_time, num=n_time_steps)
+    trajectory = solve_ivp(
+        equation_of_motion,
+        (0.0, longest_time),
+        initial_state,
+        method="RK45",
+        events=[stop_condition, z_bound_condition, rho_bound_condition],
+        args=(bfield, efield, target, particle),
+        t_eval=time_steps,
+    )
+    trajectory = trajectory.y.T
+    if len(trajectory) < 2:
+        return None
+
+    # Handle data > 90 deg. LinearInterpolator requires x data (z in our case) to be monotonically increasing
+    # So flip the array along axis 0 (flipud). Also trim data where sometimes particle starts going backward
+    # due to efield
+    if polar > np.pi * 0.5:
+        trajectory = np.flipud(trajectory)
+        mask = np.diff(np.ascontiguousarray(trajectory[:, 2])) > 0
+        mask = np.append(mask, True)
+        trajectory = trajectory[mask]
+
+    return LinearInterpolator(trajectory[:, 2], trajectory[:, :2].T)
